SCT-1454 Split e2e tests into separate files per case status type
